requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pandas-ta>=0.3.14b
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.11.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
pyyaml>=6.0
optuna>=3.0.0
joblib>=1.1.0
tqdm>=4.64.0
